Refactor ISTP codecs into istp/ subpackage
- eliminate duplication moving helpers into istp/__init__.py
- create istp/cdf.py and istp/netcdf.py
