+++ speasy/core/codecs/bundled_codecs/__init__.py
@@ -1,4 +1,4 @@
-from .istp_cdf import IstpCdf  # noqa: F401
-from .istp_netcdf import IstpNetCDF  # noqa: F401
+from .istp.cdf import IstpCdf  # noqa: F401
+from .istp.netcdf import IstpNetCDF  # noqa: F401
 from .hapi.csv import HapiCsv  # noqa: F401
 from .hapi.binary import HapiBinary  # noqa: F401
